admin * データ抽出、取込み機能実装見直し。

+++ admin/src/main/java/cherry/spring/admin/app/service/secure/userman/UsermanExportService.java
@@ -21,6 +21,6 @@ import java.util.Date;
 
 public interface UsermanExportService {
 
-	long exportUsers(Writer writer, Date registeredFrom, Date registeredTo);
+	int exportUsers(Writer writer, Date registeredFrom, Date registeredTo);
 
 }
